criado o pacote da interface GUI. close #29
