+++ README.md
@@ -1,9 +1,10 @@
 # Conversor & Divisor
 ![Conversor & Divisor](https://github.com/hadtrindade/conversor-e-divisor/workflows/Conversor%20&%20Divisor/badge.svg)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Python 3](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/python-3-shield.svg)](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/)
+[![CodeFactor](https://www.codefactor.io/repository/github/hadtrindade/conversor-e-divisor/badge)](https://www.codefactor.io/repository/github/hadtrindade/conversor-e-divisor)
+[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Updates](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/shield.svg)](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/)
-[![codecov](https://codecov.io/gh/hadtrindade/conversor-e-divisor/branch/master/graph/badge.svg?token=OMF5YWX6I6)](https://codecov.io/gh/hadtrindade/conversor-e-divisor)
+
 
 
 ## Ferramenta desenvolvida a fim de falicitar a conversão e divisão de arquivos de vídeo para o PJe, fazendo uso das ferramentas Open Souce FFMPEG e MP4Box, para converter e dividir os arquivos para o formato .mp4 e dividir arquivos convertidos suporiores a 30mb em x partes de 30mb.
