+++ ui_conversor_divisor/resources_cd_rc.py
@@ -1,6 +1,6 @@
 # Resource object code (Python 3)
 # Created by: object code
-# Created by: The Resource Compiler for Qt version 5.15.1
+# Created by: The Resource Compiler for Qt version 5.15.2
 # WARNING! All changes made in this file will be lost!
 
 from PySide2 import QtCore
@@ -1083,7 +1083,7 @@ wg\xb4\xd7\xa6\xcab\xe8Iq\xdf\xb7\x99\x5c\xa6\xa9\
 \xc1\xdfz\xb4\xe4\xfa/\xf4i\xdd\x03\x9d\xf2\xac\xff[\
 \x99\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xfc\xef\x04\x16\
 V\x09\xaf_\x22N f\
-\x00\x00\x07x\
+\x00\x00\x07\xfc\
 \x89\
 PNG\x0d\x0a\x1a\x0a\x00\x00\x00\x0dIHDR\x00\
 \x00\x00\x18\x00\x00\x00\x18\x08\x06\x00\x00\x00\xe0w=\xf8\
@@ -1140,26 +1140,26 @@ photoshop:ICCPro\
 file=\x22sRGB IEC61\
 966-2.1\x22 xmpMM:I\
 nstanceID=\x22xmp.i\
-id:798c7cb3-29cd\
--ae42-aded-e8a91\
-eb04782\x22 xmpMM:D\
+id:ef0702ba-b962\
+-ba4e-8545-0f067\
+83ad6c7\x22 xmpMM:D\
 ocumentID=\x22adobe\
 :docid:photoshop\
-:0e94af5a-4df7-6\
-849-8bb2-f527cd8\
-3b1cf\x22 xmpMM:Ori\
+:59ab4f91-fae9-1\
+648-a49a-8446708\
+70d24\x22 xmpMM:Ori\
 ginalDocumentID=\
-\x22xmp.did:eab79f8\
-5-3fa1-b64e-b245\
--528bcc2bcdc2\x22> \
+\x22xmp.did:fe4bdc4\
+2-6783-0e4e-b924\
+-6fd37ef01f3c\x22> \
 <xmpMM:History> \
 <rdf:Seq> <rdf:l\
 i stEvt:action=\x22\
 created\x22 stEvt:i\
 nstanceID=\x22xmp.i\
-id:eab79f85-3fa1\
--b64e-b245-528bc\
-c2bcdc2\x22 stEvt:w\
+id:fe4bdc42-6783\
+-0e4e-b924-6fd37\
+ef01f3c\x22 stEvt:w\
 hen=\x222020-05-02T\
 17:48:24-03:00\x22 \
 stEvt:softwareAg\
@@ -1168,10 +1168,10 @@ shop 21.0 (Windo\
 ws)\x22/> <rdf:li s\
 tEvt:action=\x22sav\
 ed\x22 stEvt:instan\
-ceID=\x22xmp.iid:79\
-8c7cb3-29cd-ae42\
--aded-e8a91eb047\
-82\x22 stEvt:when=\x22\
+ceID=\x22xmp.iid:ef\
+0702ba-b962-ba4e\
+-8545-0f06783ad6\
+c7\x22 stEvt:when=\x22\
 2020-05-02T17:52\
 :07-03:00\x22 stEvt\
 :softwareAgent=\x22\
@@ -1183,28 +1183,36 @@ stEvt:changed=\x22/\
 </rdf:Descriptio\
 n> </rdf:RDF> </\
 x:xmpmeta> <?xpa\
-cket end=\x22r\x22?>\x08q\
-\xa7o\x00\x00\x01-IDATH\xc7\xed\xd6\xbd.\
-\x84A\x14\xc6\xf1\xdf\x10\x12*\x9d\x10\x95N!\xd9\x15\
-\x91H\x08\x9d\x0b\xd0\x0b.\x00\x11\xb5P+p\x03\x94\
-\x14\x12\xbd\x86J\xeb\xa3#\x12\x89\xc4\xc4\x15\xf8H\xd8\
-W\xf3\xbe\xb2d\xf7\xdd]K\xb7O7\xc9\x93\xf3\x9f\
-sf\xe6\x9c\x09I\x92\xf8O\x85\x18#\x8c\xa3\x1f\xef\
-U|\xedx\xc5I\x8e\xa7*`\x06\xc7\xe8B\x82\x90\
-\xe3\xdf\xc0f\xa3\x805la\x15\xb7)\xa8\x5c%\xbc\
-`\x1d#\xe8I\xd7u\x03\x96\xb1\x83A\xdc\xe7x\xa7\
-q\x8a=\x1c\xa13\xc7\xdb\x89k\xdc\x84\x18\xe3\x12v\
-\xd3\xdd]T(QH\xb3h\xc3\x19&\xf1\x91\x9eK\
-%e\xde;L\x95\x03\x0a\xb8\xaa\x91\xf1\x10\x86\xd3\xb3\
-\xaav\xfd\x9e1\x8bE\x8c6\x02\x089A\x7fj!\
--e\xa1\xd1\x0cB\x1d\xa0\x04+\xd8\xfe\x0d\xa0^}\
-\xc5l\x01Z\x80\x16\xe0;\xa0\x88\xcb\x1a\xf3\xa0\x96\xb2\
-\x17\x9eu\xe8B\x881f\xcfz\x00\x8f\x7f\x94\xc1<\
-\xf6Q,\x9f\x07\x87xBG\x93\xc1\xdf0\x811\x8c\
-\x84\x18c\x01\x07\xe8K\xfb|h\x12\x90\x95\xe9\x1cs\
-\xd9\xd0\xefFo\xa3\x03=\xa7\xe3\xb6\xe1\x01\xa5\xf0\xdf\
-\xdf\x96O\x1c\x01\x9aRQ\xbc\x85+\x00\x00\x00\x00I\
-END\xaeB`\x82\
+cket end=\x22r\x22?>j!\
+`\x9b\x00\x00\x01\xb1IDATH\xc7\xd5\xd6=k\
+TA\x14\x06\xe0'\x9b\x0d\xaeh\xac\x22)\x0cV\x22\
+\xb2\x8d\x1fi\x82`TH%\xf1\x07h\xef?\xb0\xf2\
+\x0f(6b/\x82\x85\x8d\x85\x8a\x88\x8d\x88h$Q\
+D\xc4Bp\xc14\x82,\xa2\x85\x82\x9f\x89\x9bks\
+\x06\xc6\xcb\xde\xb8{a\x0b\xdf\xe6\xce\xcc\x19\xe6=\xf3\
+\x9e\x8f\xb9cEQ\x18%\x1aF\x8c\x91\x134\xbb\xdd\
+n\x13'1\x83\x1e\xca\x9aM\xe03n\xe0\xe7\xd0\x04\
+X\xc4\xcd\x01\xf6\xee\xc3\xb9:\x04{b|\x0a\xaf\xb0\
+5\xb3\x17X\xc3E\x9c\xc5e|\x18\x96`=\xc6K\
+x_\xb1\xef<N\xe0\x01\xde\xa2\x95\xd9\xc6C\xc2\xeb\
+\xb8\xdd\x8f i>\xb5\x09\xc13\x5c\xc1\x1cv\x97l\
+kX\xc0\xde*\x82A\xb0\x8e3\x9b\xd8\x1fcW\x95\
+D\xfd\xb0\x0dm\xfc\xcen\xf8+\xe6c\xd9\xbe\x8d\x98\
+Ob\x0b\x0e\xc4z\x0b\xab\xf8XEp\xe9\x1f\x1eW\
+\xe1e6~\x82\xc5~\x04;q\x1a\x8f\x22k&\xfb\
+\xd4F\x19\xbd,\xe0_#\x9d\xdb\x98\xc8\x09\xd2\xd5\xf7\
+\x87D\xd7p\xabf\x01_\xc0S|j\x94\xf4\x84c\
+\xf1]\x19\xf2\xd0\xe4\xe0\xc1\xc8\xa8\x87\xe5^\x94\xea\xe1\
+p\xa4k\xa7\xa6\xf7\xc7C\xd2\x952\xc1\xb7\xf0b\x16\
+w\x07\xd0\xbd\x8c\xb4\x7f\x01_R\xc0s\x82\x1f\xd1o\
+v\xa4\xeb\xd5\x90\xa7\x85Cq\xf8\xf72A\x13G#\
+\x16\xcfk\xca\xd3\x8e\x8ep?=\x079\xc1\xf6 \xe8\
+\xe0]My\xe6\x22U\x97\xd3z#n\xb1\x81\xe9\xd0\
+\xef^Tl\x1d\x1c\x89\x8aO\x0a\x14\xcd\xd0\xaf\x11\x0f\
+\xcaT\xb4\xed\xd9xh\x06E*\xb4y\xdcI\xfa'\
+\xdd\xdf\xe0E\xf4\x92N\x04y\xa6F\x90{x\x8d\xab\
+\x7f\x19\xfe\xfb\xbf\x8a?\xb4\xdaaG\xaae\x1f\xa5\x00\
+\x00\x00\x00IEND\xaeB`\x82\
 "
 
 qt_resource_name = b"\
@@ -1250,6 +1258,10 @@ qt_resource_name = b"\
 \x03\xc5\xb3\x93\
 \x00m\
 \x00e\x00n\x00u\x00_\x00C\x00C\
+\x00\x03\
+\x00\x00p7\
+\x00i\
+\x00m\x00g\
 \x00\x0d\
 \x0f\x0a\xde'\
 \x00c\
@@ -1283,72 +1295,81 @@ qt_resource_name = b"\
 \x00c\
 \x00i\x00l\x00-\x00m\x00e\x00d\x00i\x00a\x00-\x00s\x00t\x00o\x00p\x00.\x00p\x00n\
 \x00g\
-\x00\x0a\
-\x04R\x16E\
-\x00m\
-\x00a\x00i\x00n\x00_\x00i\x00c\x00o\x00n\x00e\
 \x00\x0e\
-\x05\x22\xe1g\
-\x00c\
-\x00i\x00l\x00-\x00f\x00o\x00l\x00d\x00e\x00r\x00.\x00p\x00n\x00g\
+\x06P\xeb\xbf\
+\x00m\
+\x00a\x00i\x00n\x00_\x00i\x00c\x00o\x00n\x00e\x00.\x00i\x00c\x00o\
 "
 
 qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x0a\x00\x00\x00\x01\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\xc0\x00\x02\x00\x00\x00\x01\x00\x00\x00\x14\
+\x00\x00\x00\xc0\x00\x02\x00\x00\x00\x01\x00\x00\x00\x1d\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x01\x08\x00\x02\x00\x00\x00\x01\x00\x00\x00\x13\
+\x00\x00\x01\x08\x00\x02\x00\x00\x00\x01\x00\x00\x00\x1b\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x12\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x19\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x002\x00\x02\x00\x00\x00\x01\x00\x00\x00\x11\
+\x00\x00\x002\x00\x02\x00\x00\x00\x01\x00\x00\x00\x17\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\xae\x00\x02\x00\x00\x00\x01\x00\x00\x00\x10\
+\x00\x00\x00\xae\x00\x02\x00\x00\x00\x01\x00\x00\x00\x15\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0f\
+\x00\x00\x00\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x13\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x8e\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0e\
+\x00\x00\x00\x8e\x00\x02\x00\x00\x00\x01\x00\x00\x00\x11\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00N\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0d\
+\x00\x00\x00N\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0f\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00x\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0c\
+\x00\x00\x00x\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0d\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00\xde\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0b\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x01<\x00\x00\x00\x00\x00\x01\x00\x00\x08\xdd\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x01\xf8\x00\x00\x00\x00\x00\x01\x00\x00(\xa2\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x01\x1c\x00\x00\x00\x00\x00\x01\x00\x00/\xbf\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x01\xbe\x00\x00\x00\x00\x00\x01\x00\x00 \xf5\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x02@\x00\x01\x00\x00\x00\x01\x00\x00?\xbd\
-\x00\x00\x01t\xd5\xf6ig\
-\x00\x00\x01\x92\x00\x00\x00\x00\x00\x01\x00\x00\x18\xb3\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x02\x16\x00\x00\x00\x00\x00\x01\x00\x008\x9c\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x02Z\x00\x00\x00\x00\x00\x01\x00\x00B\x1a\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x01\x1c\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01tr\xe9\xcc\x98\
-\x00\x00\x01h\x00\x00\x00\x00\x00\x01\x00\x00\x10\xdd\
-\x00\x00\x01tr\xe9\xcc\x98\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0c\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01H\x00\x00\x00\x00\x00\x01\x00\x00\x08\xdd\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x0e\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x02\x04\x00\x00\x00\x00\x00\x01\x00\x00(\xa2\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x10\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01(\x00\x00\x00\x00\x00\x01\x00\x00/\xbf\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x12\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01\xca\x00\x00\x00\x00\x00\x01\x00\x00 \xf5\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x14\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x02L\x00\x01\x00\x00\x00\x01\x00\x00?\xbd\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x16\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01\x9e\x00\x00\x00\x00\x00\x01\x00\x00\x18\xb3\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x18\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x02\x22\x00\x00\x00\x00\x00\x01\x00\x008\x9c\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x1a\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01H\x00\x00\x00\x00\x00\x01\x00\x00B\x1a\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x1c\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01(\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+\x00\x00\x01v3\xf10\xf4\
+\x00\x00\x01\x1c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x1e\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x01t\x00\x00\x00\x00\x00\x01\x00\x00\x10\xdd\
+\x00\x00\x01v3\xf10\xf4\
 "
 
-
 def qInitResources():
-    QtCore.qRegisterResourceData(
-        0x03, qt_resource_struct, qt_resource_name, qt_resource_data
-    )
-
+    QtCore.qRegisterResourceData(0x03, qt_resource_struct, qt_resource_name, qt_resource_data)
 
 def qCleanupResources():
-    QtCore.qUnregisterResourceData(
-        0x03, qt_resource_struct, qt_resource_name, qt_resource_data
-    )
-
+    QtCore.qUnregisterResourceData(0x03, qt_resource_struct, qt_resource_name, qt_resource_data)
 
 qInitResources()
